Finished ingredients_recipe seeder, fixed master Databaseseeder and changed ing_rec migration

+++ database/migrations/2019_01_15_094057_create_ingredient_recipe_table.php
@@ -17,7 +17,7 @@ class CreateIngredientRecipeTable extends Migration
             $table->increments('id');
             $table->integer('ingredient_id')->unsigned();
             $table->integer('recipe_id')->unsigned();
-            $table->integer('amount');
+            $table->decimal('amount',8,3);
             $table->integer('unit_id')->unsigned();
 
             $table->foreign('ingredient_id')->references('id')->on('ingredients');;
